fix(adr-061): break model.source_graph <-> buildsource import cycle
A Codex review on PR #937 caught a real circular-import failure the
previous commit introduced: `import abicheck.model.source_graph` as a
first touch raised ImportError, because importing any submodule of
abicheck.buildsource (entity_resolver.py, graph_facts.py) first runs
buildsource/__init__.py, which eagerly imports call_graph.py, which
imports the legacy buildsource/source_graph.py facade, which imports
back from abicheck.model.source_graph -- still mid-initialization.
Existing tests never hit this because they all reach source_graph
through the legacy facade first.

Fixed at the root: graph_facts.py and the EntityResolver/entity_identity
machinery physically move into abicheck/model/ (none of them depend on
buildsource at all, only abicheck.name_classification/abicheck.demangle),
so model/source_graph.py's own imports become same-package relative and
never touch the abicheck.buildsource namespace. The three old buildsource/
paths become thin `X as X` re-export facades so every existing same-package
relative import and absolute import (used directly by several tests) keeps
resolving.

graph_facts.py (1123 lines, moved unchanged) split three ways to clear the
new-file 800-line cap the move triggered: graph_vocabulary.py (confidence
labels + node/edge-kind vocabulary), graph_identity.py (decl/type
id-normalization -- already split out of source_graph.py once before for
the identical line-cap reason), and graph_facts.py itself (GraphFact/
GraphNode/GraphEdge/merge machinery).

entity_identity.py's `from .. import demangle` became `from ..demangle
import demangle as _demangle` so check_architecture.py's import scan
resolves one unambiguous target instead of an ambiguous bare-package
reference. demangle.py and entity_resolver.py join architecture/
modules.yaml's model legacy_paths; the now-stale graph_facts.py/
entity_resolver.py legacy_paths entries (from the previous commit, before
this one physically relocated them) are removed.

Verified: import abicheck.model.source_graph (and every other entry
point) now succeeds on a fresh interpreter; check_architecture.py 0
errors; mypy clean; ruff clean; every test file referencing any moved
name directly (767 tests) passes unchanged; full fast suite re-run clean.
