+++ abicheck/model/AGENTS.md
@@ -50,6 +50,10 @@ types, so `from abicheck.elf_metadata import ElfMetadata` still resolves.
 | A new ELF/PE/Mach-O/DWARF/SYCL/kABI fact | the matching `*_facts.py` |
 | A snapshot-aware surface predicate | `stdlib_surface.py` |
 | An L5 source-graph value field, node-id spelling, or schema vocabulary entry (`NODE_KINDS`/`EDGE_KINDS`) | `source_graph.py` |
+| An L5 `GraphNode`/`GraphEdge` field, or the ADR-046 fact-merge machinery | `graph_facts.py` |
+| An L5 confidence label or `*_NODE_KINDS`/`*_EDGE_KINDS` family vocabulary set | `graph_vocabulary.py` |
+| A decl/type node-id normalization rule | `graph_identity.py` |
+| A `EntityResolver`/canonical-identity resolution rule | `entity_resolver.py`, `entity_identity.py` |
 
 `__init__.py` is the supported import surface and stays a re-export list
 with `__all__` — no logic, no new names that are not owned by a submodule.
